Move classes for consistency
