+++ app-modules/case-management/src/Notifications/EducatableCaseClosedNotification.php
@@ -40,7 +40,7 @@ use AdvisingApp\CaseManagement\Models\CaseModel;
 use AdvisingApp\Notification\Enums\NotificationChannel;
 use AdvisingApp\Notification\Models\Contracts\CanBeNotified;
 use AdvisingApp\Notification\Models\OutboundDeliverable;
-use AdvisingApp\Notification\Notifications\HasBeforeSendHook;
+use AdvisingApp\Notification\Notifications\Contracts\HasBeforeSendHook;
 use AdvisingApp\Notification\Notifications\Messages\MailMessage;
 use AdvisingApp\Prospect\Models\Prospect;
 use AdvisingApp\StudentDataModel\Models\Contracts\Educatable;
@@ -51,7 +51,7 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\AnonymousNotifiable;
 use Illuminate\Notifications\Notification;
 
-class SendEducatableCaseOpenedNotification extends Notification implements ShouldQueue, HasBeforeSendHook
+class EducatableCaseClosedNotification extends Notification implements ShouldQueue, HasBeforeSendHook
 {
     use Queueable;
 
@@ -78,15 +78,12 @@ class SendEducatableCaseOpenedNotification extends Notification implements Shoul
         };
 
         $status = $this->case->status;
-        $type = $this->case->priority->type;
 
         return MailMessage::make()
             ->settings($this->resolveNotificationSetting($notifiable))
             ->subject("{$this->case->case_number} - is now {$status->name}")
-            ->greeting("Hello {$name},")
-            ->line("A new {$type->name} case has been created and is now in a {$status->name} status. Your new ticket number is: {$this->case->case_number}.")
-            ->line('The details of your case are shown below:')
-            ->lines(str(nl2br($this->case->close_details))->explode('<br />'));
+            ->greeting("Hi {$name},")
+            ->line("Your request {$this->case->case_number} for case is now {$status->name}.");
     }
 
     public function beforeSend(AnonymousNotifiable|CanBeNotified $notifiable, OutboundDeliverable $deliverable, NotificationChannel $channel): void
