+++ app-modules/case-management/src/Notifications/EducatableCaseOpenedNotification.php
@@ -34,24 +34,29 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\MeetingCenter\Notifications;
+namespace AdvisingApp\CaseManagement\Notifications;
 
-use AdvisingApp\MeetingCenter\Models\Event;
-use AdvisingApp\MeetingCenter\Models\EventAttendee;
+use AdvisingApp\CaseManagement\Models\CaseModel;
+use AdvisingApp\Notification\Enums\NotificationChannel;
+use AdvisingApp\Notification\Models\Contracts\CanBeNotified;
+use AdvisingApp\Notification\Models\OutboundDeliverable;
+use AdvisingApp\Notification\Notifications\Contracts\HasBeforeSendHook;
 use AdvisingApp\Notification\Notifications\Messages\MailMessage;
+use AdvisingApp\Prospect\Models\Prospect;
+use AdvisingApp\StudentDataModel\Models\Contracts\Educatable;
+use AdvisingApp\StudentDataModel\Models\Student;
 use App\Models\NotificationSetting;
-use App\Models\User;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Notifications\AnonymousNotifiable;
 use Illuminate\Notifications\Notification;
 
-class SendRegistrationLinkToEventAttendeeNotification extends Notification implements ShouldQueue
+class EducatableCaseOpenedNotification extends Notification implements ShouldQueue, HasBeforeSendHook
 {
     use Queueable;
 
     public function __construct(
-        protected Event $event,
-        protected User $sender
+        protected CaseModel $case,
     ) {}
 
     /**
@@ -64,17 +69,33 @@ class SendRegistrationLinkToEventAttendeeNotification extends Notification imple
 
     public function toMail(object $notifiable): MailMessage
     {
+        /** @var Educatable $educatable */
+        $educatable = $notifiable;
+
+        $name = match ($notifiable::class) {
+            Student::class => $educatable->first,
+            Prospect::class => $educatable->first_name,
+        };
+
+        $status = $this->case->status;
+        $type = $this->case->priority->type;
+
         return MailMessage::make()
             ->settings($this->resolveNotificationSetting($notifiable))
-            ->subject('You have been invited to an event!')
-            ->line("You have been invited to {$this->event->title}.")
-            ->action('Register', route('event-registration.show', ['event' => $this->event]));
+            ->subject("{$this->case->case_number} - is now {$status->name}")
+            ->greeting("Hello {$name},")
+            ->line("A new {$type->name} case has been created and is now in a {$status->name} status. Your new ticket number is: {$this->case->case_number}.")
+            ->line('The details of your case are shown below:')
+            ->lines(str(nl2br($this->case->close_details))->explode('<br />'));
+    }
+
+    public function beforeSend(AnonymousNotifiable|CanBeNotified $notifiable, OutboundDeliverable $deliverable, NotificationChannel $channel): void
+    {
+        $deliverable->related()->associate($this->case);
     }
 
     private function resolveNotificationSetting(object $notifiable): ?NotificationSetting
     {
-        return $notifiable instanceof EventAttendee
-            ? $this->sender->teams()->first()?->division?->notificationSetting?->setting
-            : null;
+        return $this->case->division?->notificationSetting?->setting;
     }
 }
