+++ app-modules/notification/src/Notifications/Channels/DatabaseChannel.php
@@ -40,8 +40,8 @@ use AdvisingApp\Notification\Actions\MakeOutboundDeliverable;
 use AdvisingApp\Notification\DataTransferObjects\DatabaseChannelResultData;
 use AdvisingApp\Notification\Enums\NotificationChannel;
 use AdvisingApp\Notification\Enums\NotificationDeliveryStatus;
-use AdvisingApp\Notification\Notifications\HasAfterSendHook;
-use AdvisingApp\Notification\Notifications\HasBeforeSendHook;
+use AdvisingApp\Notification\Notifications\Contracts\HasAfterSendHook;
+use AdvisingApp\Notification\Notifications\Contracts\HasBeforeSendHook;
 use Illuminate\Notifications\Channels\DatabaseChannel as BaseDatabaseChannel;
 use Illuminate\Notifications\Notification;
 use Throwable;
